v0.53.0: DX, Extended Standards & Architecture
## Summary

This commit delivers v0.53.0 of pg_ripple. All roadmap deliverables are
implemented, code is clippy-clean, and tests pass.

## New features

### SHACL-SPARQL sh:SPARQLConstraintComponent (N5-3 / F-9)
- New `SparqlConstraint` variant on `ShapeConstraint` enum
- Parser recognises `sh:sparql` predicates in node/property shapes
- `dispatch_constraint()` executes the SPARQL SELECT query with `$this`
  bound to the focus-node IRI; any non-empty result set → violation
- New src/shacl/constraints/sparql_constraint.rs with
  `check_sparql_constraint()`
- pg_regress test: tests/pg_regress/sql/shacl_sparql_constraint.sql

### pg_ripple.copy_rdf_from(path, format) (N9-5 / F-10)
- New SQL function dispatching to load_ntriples_file / load_nquads_file /
  load_turtle_file / load_trig_file / load_rdfxml_file based on format
- Added to src/dict_api.rs in the pg_ripple schema block

### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAon ### RAGon_ha### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG piw ### RAG pi### RAG pi### RAG pi### RAG pi### RAG prat### RAG pi### # C### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG e### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi#op### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAdde### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pets (N4-### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG pi### RAG l
- fuzz/fuzz_targets/jsonld_framer.rs — JSON-LD via serde_json
- fuzz/fuzz_targets/ht- fuzz/fuzz_targets/ht- fuzz/fuzz_targets/ht- fuzz/fuzz_tarependencies rio_xml, serde_json, url added to fuzz/Cargo.toml

###########################################################################di###################################################################2-3)#######################################################at ###########################################################################di###################################################################2-3)#######################################################at ###########################################################################di###################################################################2-3)#######################################################at ###########################################################################di###################################################################2-3)################################################src/datalog/seminaive.rs — semi-naive evaluator delegation

### HTTP unwrap() hardening (N1-3)
- pg_ripple_http/src/main.rs: 10 Response::builder().unwrap() →
  unwrap_or_else(|e| structured 500 response)

## Version bumps
- Cargo.toml: 0.52.0 → 0.53.0
- pg_ripple.control: 0.52.0 → 0.53.0
- Migration script: sql/pg_ripple--0.52.0--0.53.0.sql
