+++ .github/workflows/ci.yml
@@ -489,10 +489,10 @@ jobs:
         run: cargo pgrx stop pg18
 
   watdiv-suite:
-    name: WatDiv benchmark suite (non-blocking)
+    name: WatDiv benchmark suite
     runs-on: ubuntu-latest
     needs: [test, w3c-suite]
-    continue-on-error: true
+    continue-on-error: false
 
     steps:
       - name: Checkout
