+++ clippy.log
@@ -1,0 +1,2 @@
+   Compiling pg_ripple v0.52.0 (/Users/geir.gronmo/projects/pg-ripple2)
+    Finished `dev` profile [unoptimized + debuginfo] target(s) in 2.32s
