+++ roadmap/v0.53.0-full.md
@@ -15,32 +15,32 @@
 
 #### Standards Completeness
 
-- [ ] **SHACL-SPARQL `sh:SPARQLConstraintComponent`** (N5-3 / F-9): W3C SHACL-SPARQL; `sh:SPARQLConstraintComponent` executes a user-authored SPARQL query as a constraint; results decoded into `Violation` structs; add pg_regress `shacl_sparql_constraint.sql`
-- [ ] **SHACL-AF `sh:rule` warning / compile** (S4-8 / N9-3): emit **PT480** `ShAFRuleUnsupported` when `sh:rule` triples are detected; or (preferred path) compile SPARQLRules to the existing Datalog rule engine
-- [ ] **`COPY rdf FROM` integration** (N9-5 / F-10): register a custom PostgreSQL `COPY` handler so `COPY pg_ripple.triples FROM '/path/to.nt' WITH (FORMAT 'ntriples')` and `(FORMAT 'turtle')` work as first-class PostgreSQL commands with batched dictionary encoding; return row count
+- [x] **SHACL-SPARQL `sh:SPARQLConstraintComponent`** (N5-3 / F-9): W3C SHACL-SPARQL; `sh:SPARQLConstraintComponent` executes a user-authored SPARQL query as a constraint; results decoded into `Violation` structs; add pg_regress `shacl_sparql_constraint.sql`
+- [x] **SHACL-AF `sh:rule` warning / compile** (S4-8 / N9-3): emit **PT480** `ShAFRuleUnsupported` when `sh:rule` triples are detected; or (preferred path) compile SPARQLRules to the existing Datalog rule engine
+- [x] **`COPY rdf FROM` integration** (N9-5 / F-10): register a custom PostgreSQL `COPY` handler so `COPY pg_ripple.triples FROM '/path/to.nt' WITH (FORMAT 'ntriples')` and `(FORMAT 'turtle')` work as first-class PostgreSQL commands with batched dictionary encoding; return row count
 
 #### RAG & LLM Hardening
 
-- [ ] **RAG pipeline hardening** (F-11): sanitise NL input before LLM call (block prompt-injection patterns; truncate at `llm_max_input_tokens`); add response caching keyed on `(question_hash, k, schema_digest)` in `_pg_ripple.rag_cache`; add `pg_ripple_http /rag` REST endpoint; add `examples/llm_workflow.sql` if not added in v0.51.0
+- [x] **RAG pipeline hardening** (F-11): sanitise NL input before LLM call (block prompt-injection patterns; truncate at `llm_max_input_tokens`); add response caching keyed on `(question_hash, k, schema_digest)` in `_pg_ripple.rag_cache`; add `pg_ripple_http /rag` REST endpoint; add `examples/llm_workflow.sql` if not added in v0.51.0
 
 #### CDC & Subscriptions
 
-- [ ] **CDC lifecycle events** (N9-2): second NOTIFY channel `pg_ripple_cdc_lifecycle_{name}` emitting merge-cycle events (`{"op":"merge","predicate_id":N,"merged":M,"tombstones":T}`) and VP promotion events (`{"op":"promote","predicate_id":N,"from":"rare","to":"vp"}`)
+- [x] **CDC lifecycle events** (N9-2): second NOTIFY channel `pg_ripple_cdc_lifecycle_{name}` emitting merge-cycle events (`{"op":"merge","predicate_id":N,"merged":M,"tombstones":T}`) and VP promotion events (`{"op":"promote","predicate_id":N,"from":"rare","to":"vp"}`)
 
 #### Test Coverage
 
 - [ ] **Property-based test generator enrichment** (N4-2): enrich dictionary generator (NFC/NFD Unicode, RTL, emoji, zero-width), SPARQL round-trip generator (property paths, subqueries, aggregates), JSON-LD framing generator (nested `@context`, `@list`, `@container`)
-- [ ] **RDF/XML + JSON-LD framing fuzz targets** (N4-3): `fuzz/fuzz_targets/rdfxml_parser.rs` and `fuzz/fuzz_targets/jsonld_framer.rs`; assert no panic on arbitrary input
-- [ ] **HTTP companion fuzz coverage** (N4-4): `fuzz/fuzz_targets/http_request.rs` targeting the `pg_ripple_http` request-handler chain
-- [ ] **WatDiv conformance gate flip** (N4-5): promote WatDiv latency baseline gate from warning to blocking after two consecutive stable releases
+- [x] **RDF/XML + JSON-LD framing fuzz targets** (N4-3): `fuzz/fuzz_targets/rdfxml_parser.rs` and `fuzz/fuzz_targets/jsonld_framer.rs`; assert no panic on arbitrary input
+- [x] **HTTP companion fuzz coverage** (N4-4): `fuzz/fuzz_targets/http_request.rs` targeting the `pg_ripple_http` request-handler chain
+- [x] **WatDiv conformance gate flip** (N4-5): promote WatDiv latency baseline gate from warning to blocking after two consecutive stable releases
 
 #### Architecture & Code Quality
 
-- [ ] **`gucs.rs` subsystem split** (N1-1): split 1,617-line `src/gucs.rs` into `src/gucs/{sparql.rs, datalog.rs, shacl.rs, federation.rs, llm.rs, storage.rs, observability.rs}`
-- [ ] **`src/datalog/mod.rs` split** (N1-4): extract semi-naïve evaluator → `seminaive.rs`, magic-set transformer → `magic.rs`, demand-filter rewrite → `demand.rs`, parallel-strata coordinator → `coordinator.rs`
-- [ ] **`filter.rs` split** (N7-1): split 901-line `src/sparql/translate/filter.rs` into `filter_expr.rs` (expression compilation) and `filter_dispatch.rs` (pattern dispatch)
-- [ ] **HTTP companion hot-path `unwrap()` fixes** (N1-3): convert request-handler-path `unwrap()` calls at [pg_ripple_http/src/main.rs:829,865](pg_ripple_http/src/main.rs#L829) to `?`-propagation with HTTP 400/500 responses
-- [ ] **Merge-throughput baseline** (N2-3): record p50/p95 at `merge_workers ∈ {1,2,4,8}` in `benchmarks/merge_throughput_baselines.json`; add CI warning gate on >15 % regression
+- [x] **`gucs.rs` subsystem split** (N1-1): split 1,617-line `src/gucs.rs` into `src/gucs/{sparql.rs, datalog.rs, shacl.rs, federation.rs, llm.rs, storage.rs, observability.rs}`
+- [x] **`src/datalog/mod.rs` split** (N1-4): extract semi-naïve evaluator → `seminaive.rs`, magic-set transformer → `magic.rs`, demand-filter rewrite → `demand.rs`, parallel-strata coordinator → `coordinator.rs`
+- [x] **`filter.rs` split** (N7-1): split 901-line `src/sparql/translate/filter.rs` into `filter_expr.rs` (expression compilation) and `filter_dispatch.rs` (pattern dispatch)
+- [x] **HTTP companion hot-path `unwrap()` fixes** (N1-3): convert request-handler-path `unwrap()` calls at [pg_ripple_http/src/main.rs:829,865](pg_ripple_http/src/main.rs#L829) to `?`-propagation with HTTP 400/500 responses
+- [x] **Merge-throughput baseline** (N2-3): record p50/p95 at `merge_workers ∈ {1,2,4,8}` in `benchmarks/merge_throughput_baselines.json`; add CI warning gate on >15 % regression
 - [ ] **HTAP merge cutover atomic fix** (C-3): investigate and, if confirmed safe, eliminate the `CREATE OR REPLACE VIEW` step in [src/storage/merge.rs:331-346](src/storage/merge.rs#L331); add concurrent-merge stress test asserting zero `relation does not exist` errors over 50 parallel queries
 
 ### Migration Script
