+++ src/shacl/constraints/mod.rs
@@ -8,6 +8,7 @@ pub mod logical;
 pub mod property_path;
 pub mod relational;
 pub mod shape_based;
+pub mod sparql_constraint;
 pub mod string_based;
 pub mod value_type;
