+++ tests/pg_regress/expected/test.txt
@@ -1,0 +1,1 @@
+test
